drupal7: more changes to install script

+++ scripts/DRUPAL7/00_install.sh
@@ -8,10 +8,15 @@ $ROOTCMD curl -L https://ftp.drupal.org/files/projects/drupal-$ver.tar.gz -o /tm
 $ROOTCMD mkdir -p $base
 $ROOTCMD tar xzf /tmp/drupal-$ver.tar.gz -C$base/
 $ROOTCMD rm -f /tmp/drupal-$ver.tar.gz
+$ROOTCMD ls -l $base
 $ROOTCMD mv $base/drupal-$ver/* $base/drupal-$ver/.htaccess $base/drupal-$ver/.gitignore $base
-$ROOTCMD cp $base/sites/default/default.settings.php $base/sites/default/settings.php
-$ROOTCMD chown $user $base/sites/default/settings.php
-$ROOTCMD mkdir $base/sites/default/files
-$ROOTCMD chown $user $base/sites/default/files
+if [ ! -f $base/sites/default/settings.php ]; then
+	$ROOTCMD cp $base/sites/default/default.settings.php $base/sites/default/settings.php
+	$ROOTCMD chown $user $base/sites/default/settings.php
+fi
+if [ ! -d $base/sites/default/files ]; then
+	$ROOTCMD mkdir $base/sites/default/files
+	$ROOTCMD chown $user $base/sites/default/files
+fi
 $ROOTCMD rm -rf $base/drupal-$ver
 $ROOTCMD rm -rf $base/html
